refactor: move model canonicalization into lib/prismic/models
