+++ src/commands/pull.ts
@@ -1,11 +1,11 @@
 import { getAdapter } from "../adapters";
 import { getCredentials } from "../auth";
-import { canonicalizeCustomType, canonicalizeSlice } from "../canonicalize";
 import { CommandError, createCommand, type CommandConfig } from "../lib/command";
 import { diffArrays } from "../lib/diff";
 import { getDirtyPaths, getGitRoot } from "../lib/git";
 import { getCustomTypes, getSlices } from "../lib/prismic/clients/custom-types";
 import { completeOnboardingStepsSilently } from "../lib/prismic/clients/repository";
+import { canonicalizeCustomType, canonicalizeSlice } from "../lib/prismic/models";
 import { isDescendant, relativePathname } from "../lib/url";
 import { findProjectRoot, getRepositoryName } from "../project";
