+++ src/commands/push.ts
@@ -4,7 +4,6 @@ import { pascalCase } from "change-case";
 
 import { getAdapter } from "../adapters";
 import { getCredentials } from "../auth";
-import { canonicalizeCustomType, canonicalizeSlice } from "../canonicalize";
 import { CommandError, createCommand, type CommandConfig } from "../lib/command";
 import { diffArrays } from "../lib/diff";
 import { getDirtyPaths, getGitRoot } from "../lib/git";
@@ -24,6 +23,7 @@ import {
 	completeOnboardingStepsSilently,
 	type OnboardingStep,
 } from "../lib/prismic/clients/repository";
+import { canonicalizeCustomType, canonicalizeSlice } from "../lib/prismic/models";
 import { BadRequestError } from "../lib/request";
 import { appendTrailingSlash, isDescendant, relativePathname } from "../lib/url";
 import { findProjectRoot, getRepositoryName } from "../project";
